Стварыць асобны файл для канвертэра Кірыліца->Лацінка
